adjust cli
