+++ cli/cmd/root.go
@@ -20,11 +20,13 @@ package cmd
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/spacechunks/explorer/cli"
 	"github.com/spacechunks/explorer/cli/cmd/register"
 	"github.com/spacechunks/explorer/cli/cmd/version"
 	"github.com/spf13/cobra"
+	"google.golang.org/grpc/metadata"
 )
 
 func Root(ctx context.Context, cliCtx cli.Context) *cobra.Command {
@@ -42,4 +44,24 @@ A place of discovery and play. All within a single unified system.`,
 	)
 
 	return root
+}
+
+func requireAccessToken(
+	ctx context.Context,
+	cliCtx cli.Context,
+	fn func(context.Context, cli.Context) *cobra.Command,
+) *cobra.Command {
+	cmd := fn(ctx, cliCtx)
+	cmd.RunE = func(cmd *cobra.Command, args []string) error {
+		tok, err := cliCtx.Auth.AccessToken(ctx)
+		if err != nil {
+			return fmt.Errorf("authentication failed: %w", err)
+		}
+
+		md := metadata.Pairs("authorization", tok)
+		ctx = metadata.NewOutgoingContext(ctx, md)
+
+		return fn(ctx, cliCtx).RunE(cmd, args)
+	}
+	return cmd
 }
